==> 02_add_url_and_rank.ipynb <==
### This workbook returns a website URL and associated Google Search rank for a list of input keywords using the serper.dev Google Search API

import requests
import json
import pandas as pd
from typing import List, Optional
pd.set_option('display.max_rows', 100)
pd.set_option('display.max_columns', 10)

### Functions

def load_api_key(key_path: str='secret/key.txt', key_name: str='api_key') -> None:
    key_name = str(key_name)
    with open(key_path, 'r') as f:
        key_value = f.read().strip()
    globals()[key_name] = key_value

    print(f"API key set to global varible '{key_name}'")

def search_kw(kw: str, page: int = 1, url: str = "https://google.serper.dev/search") -> requests.Response:
    """
    Searches for a given keyword using the Serper API and returns the response object.

    Args:
        kw (str): The keyword to search for.
        page (int, optional): The page number of the search results. Defaults to 1.
        url (str, optional): The serper.dev API endpoint. Defaults to "https://google.serper.dev/search".

    Returns:
        requests.Response: The response object containing the search results.
    """

    payload = json.dumps({
      "q": kw, # keyword to search for
      "gl": "us", # location
      "hl": "en", # language
      "autocorrect": False,
      "page": page
    })

    headers = {
      "X-API-KEY": api_key,
      "Content-Type": "application/json"
    }

    response = requests.request("POST", url=url, headers=headers, data=payload)
    return response

def print_response(response: requests.Response) -> None:
    """
    Prints the search results from the serper.dev API response object.

    Args:
        response (requests.Response): The response object containing the search results.

    Returns:
        None
    """

    search_term = response.json()['searchParameters']['q']
    print(f"Results for keyword: \"{search_term}\"")
    
    for rank, result in enumerate(response.json()["organic"], 1):
        title = result['title']
        link = result['link']
        print(f"Rank: {rank}, Title: {title}, Link: {link}")

def get_pages(page_count: int, kw: str, url: str = "https://google.serper.dev/search") -> List[requests.Response]:
    """
    Retrieves the specified number of pages of search results for a given keyword using the Serper API.

    Args:
        page_count (int): The number of pages of search results to retrieve.
        kw (str): The keyword to search for.
        url (str, optional): The Serper API endpoint. Defaults to "https://google.serper.dev/search".

    Returns:
        List[requests.Response]: A list of response objects, where each object contains the search results for a single page.
    """

    pages = []
    for page in range(1, page_count+1):
        pages.append(search_kw(kw, page, url))
    return pages

def extract_response(response: requests.Response) -> pd.DataFrame:
    """
    Extracts the search results from a Serper API response object and returns them as a pandas DataFrame.

    Args:
        response (requests.Response): The response object containing the search results.

    Returns:
        pd.DataFrame: A DataFrame containing the search results, with columns for the search term (kw), 
        page number (page), rank of the result on the page (rank), and the link to the result (link).
    """

    df = pd.DataFrame(columns = ['kw', 'page', 'rank', 'link'])
    q = response.json()['searchParameters']['q']
    page = response.json()['searchParameters']['page']
    for rank, result in enumerate(response.json()["organic"], 1):
        link = result['link']
        df.loc[len(df)] = [q, page, rank, link]
    return df

def collate_pages(pages: List[requests.Response], max_rank: Optional[int] = None) -> pd.DataFrame:
    """
    Collates the search results from a list of serper.decv API response objects and returns them as a pandas DataFrame.

    Args:
        pages (List[requests.Response]): A list of response objects, where each object contains the search results for a single page.
        max_rank (int, optional): The maximum rank of search results to include in the output DataFrame. Defaults to None.

    Returns:
        pd.DataFrame: A DataFrame containing the collated search results, with columns for the search term (kw), 
        rank of the result (rank), and the link to the result (link).
    """

    df = pd.DataFrame()
    for response in pages:
        data = extract_response(response)
        df = pd.concat([df, data], ignore_index=True)
    df = df.reset_index(drop=True)
    df["rank"] = df.index + 1
    if max_rank is not None:
        df = df[df['rank'] <= max_rank]
    df.drop(["page"], axis=1, inplace=True)
    return df

def get_links_for_kws(df_in: pd.DataFrame) -> pd.DataFrame:
    """
    Gets search results for a list of keywords and returns them as a pandas DataFrame.

    Args:
        df_in (pd.DataFrame): A DataFrame containing a column of keywords under the column name "Keyword".

    Returns:
        pd.DataFrame: A DataFrame containing the collated search results for each keyword, with columns for the 
        search term (kw), rank of the result (rank), and the link to the result (link).
    """

    df_out = pd.DataFrame()
    keywords = df_in["kw"]
    for kw in keywords:
        kw_pages = get_pages(3, kw) # get 3 pages (<= 30 records) for each kw
        df = collate_pages(kw_pages, 20) # collate pages, limit to top 20 records.
        df_out = pd.concat([df_out, df], ignore_index=True)
    return(df_out)

def combine_csvs_in_df(files: List[str]) -> pd.DataFrame:
    """
    Combines multiple CSV files into a single DataFrame.

    Args:
        files_to_combine (List[str]): A list of file paths for the CSV files to be combined.

    Returns:
        pd.DataFrame: A DataFrame containing the combined data from all the input CSV files.
    """

    combined_df = pd.DataFrame()
    for file in files:
        df = pd.read_csv(file, header=0)
        combined_df = pd.concat([combined_df, df], ignore_index=True)
    return combined_df

def print_dfs_lengths(df_names: List[str]) -> int:
    """
    Prints the length of each DataFrame in a list of DataFrame names and returns the total length.

    Args:
        df_names (List[str]): A list of strings representing the names of the DataFrames to be printed.

    Returns:
        int: The total length of all the DataFrames.
    """

    # Get the length of each DataFrame
    try:
        lengths = [len(globals()[df_name]) for df_name in df_names]
        
        # Print the length of each DataFrame
        for i, length in enumerate(lengths):
            print(f"DataFrame {i+1} ({df_names[i]}): {length}")
        
        # Calculate and print the total length of all the DataFrames
        total_length = sum(lengths)
        print(f"Total length: {total_length}")
        
        return total_length
    except:
        print("There was a problem")

### Code Execution Examples

load_api_key(key_path='secret/key.txt', key_name='api_key')

# load data/kw_1_000.csv

df_in = pd.read_csv('data/kw_1_000.csv')
print(f"df_in has {len(df_in)} records")
df_in.head(3)

# example of search for 1 kw and 1 page
response = search_kw(kw='space', page=1)
response

print_response(response)

#extract response into df

extract_response(response)

response2 = get_pages(page_count=2, kw='nasa')
response2

collate_pages(response2)

df_in[:1]

# attempts to get 20 links per kw

response3 = get_links_for_kws(df_in[:1])
response3

df_2 = df_in.iloc[0:2]
df_linked_kw_out_2 = get_links_for_kws(df_2) # DOES 6 QUERIES (3 pages for each)
df_linked_kw_out_2

### Code Execution

load_api_key(key_path='secret/key.txt', key_name='api_key')

# load data/kw_1_000.csv

df_in = pd.read_csv('data/kw_1_000.csv')
print(f"df_in has {len(df_in)} records")
df_in.head(3)

# spliting the data into batches of 400 or less
df_400 = df_in.iloc[0:400]
df_400_800 = df_in.iloc[400:800]
df_800_1000 = df_in.iloc[800:]

df_400_800.head()

# ***DANGER*** GETS DATA FROM SERPER.API
# REMEMBER YOU ARE LIMITED BY A DAILY QUERY QUOTA OF 1440
# UNCOMMENT TO RUN

#df_linked_kw_out_400 = get_links_for_kws(df_400) # DOES 1200 QUERIES (3 pages for each)
#df_linked_kw_out_400_800 = get_links_for_kws(df_400_800) # DOES 1200 QUERIES (3 pages for each)
#df_linked_kw_out_800_1000 = get_links_for_kws(df_800_1000) # DOES 600 QUERIES (3 pages for each)

df_names = ['df_linked_kw_400', 'df_linked_kw_400_800', 'df_linked_kw_800_1000']
print_dfs_lengths(df_names)

# SAVE LINKED KW DFs TO CSV 
df_linked_kw_400.to_csv('data/linked_kw_400.csv', index=False)
df_linked_kw_400_800.to_csv('data/linked_kw_400_800.csv', index=False)
df_linked_kw_800_1000.to_csv('data/linked_kw_800_1000.csv', index=False)

files = ['data/linked_kw_400.csv', 'data/linked_kw_400_800.csv', 'data/linked_kw_800_1000.csv']
files

df_linked_kw_final = combine_csvs_in_df(files)
df_linked_kw_final.tail()

# save to csv
df_linked_kw_final.to_csv('data/linked_kw_final.csv', index=False)

# reload from csv
df_linked_kw_final = pd.read_csv('data/linked_kw_final.csv', header=0)
print(f"df_linked_kw_final has {len(df_linked_kw_final)} records")
df_linked_kw_final.head(3)

len(df_linked_kw_final)

df_linked_kw_final.tail()

kw_counts = df_linked_kw_final['kw'].value_counts()
kw_less_than_20 = kw_counts[kw_counts < 20].index.tolist()

# this is the one with only ten links

print(kw_less_than_20)
