==> 04_batch_reassembly.ipynb <==
### This notebook used to

import pandas as pd
import numpy as np
import os
from typing import List, Dict, Optional
import inspect
pd.set_option('display.max_rows', 400)
pd.set_option('display.max_columns', 100)

### Functions

def show_df_info(df: pd.DataFrame) -> pd.DataFrame:
    """
    Prints information about a DataFrame, including column names, data types, and non-null counts.
    
    Args:
        df (pandas.DataFrame): The DataFrame to process.
        
    Returns:
        pandas.DataFrame: A DataFrame containing the column names, data types, and non-null counts, and null counts of the input DataFrame.
    """
    col_names = df.columns.to_list()
    col_dtypes = df.dtypes.to_list()
    non_null_counts = df.count().to_list()
    null_counts = df.isnull().sum().to_list()
    info_df = pd.DataFrame({'column_name': col_names, 'dtype': col_dtypes, 'non_null_count': non_null_counts, 'null_count': null_counts})

    caller_frame = inspect.currentframe().f_back
    df_name = [var_name for var_name, var_val in caller_frame.f_locals.items() if var_val is df][0]

    print(f"DataFrame '{df_name}' has {len(df)} rows and {len(df.columns)} columns.")
    print("Here is a summary of the column names, data types and null counts:")
    return info_df

def assemble_csv(in_dir: str) -> pd.DataFrame:
    """
    Reads all CSV files from the specified directory and loads them into a Pandas DataFrame.

    Args:
        in_dir (str): The input directory containing the CSV files.

    Returns:
        A Pandas DataFrame containing the concatenated data from all CSV files.
    """
    # Create an empty list to store the data from all CSV files
    df_list = list()

    # Iterate over all files in the input directory
    for filename in os.listdir(in_dir):
        if filename.endswith('.csv'):
            # Read the CSV file into a Pandas DataFrame
            df = pd.read_csv(os.path.join(in_dir, filename))

            # Append the data frame to the list
            df_list.append(df)

    # Concatenate all data frames into a single data frame
    combined_df = pd.concat(df_list)

    # Return the combined data frame
    return combined_df

in_dir = 'data/batches_classified'
df = assemble_csv(in_dir)
df.head(1)

#df['uid'] = df['kw'] + '*' + df['link']

df['uid'] = df.apply(lambda row: f"{row['kw']}_{row['rank']}_{row['link']}", axis=1)

show_df_info(df)

df.head(1)

len(df['link'])

len(df['link'][0])

#==df['link'][0]

df = df.drop(['link.1'], axis=1)

df.head()

non_null_count = df['word_count'].count()
print("Number of non-null values in 'word_count':", non_null_count)

true_count = df['success'].sum()
print("Number of True values in 'success':", true_count)

true_count = (df['success'] == False).sum()
true_percent = true_count / len(df) * 100
true_percent

class_zero = ((df['percent_human']== 0) & (df['percent_ai'] == 0)).sum()
class_zero

class_zero = ((df['percent_human']== 0) & (df['percent_ai'] == 0)).sum()
class_nan = ((df['percent_human'].isna()) | (df['percent_ai'].isna())).sum()
x = len(df) - class_zero - class_nan
x

class_nonzero = ((df['percent_human']== 0) & (df['percent_ai'] == 0) == False).sum()
class_nonzero

def summarize_results(df):
    success_true = (df['success'] == True).sum()
    success_false = (df['success'] == False).sum()
    class_zero = ((df['percent_human']== 0) & (df['percent_ai'] == 0)).sum()
    class_nonzero = ((df['percent_human']!= 0) & (df['percent_ai'] != 0)).sum()
    class_nan = ((df['percent_human'].isna()) | (df['percent_ai'].isna())).sum()
    print(f'success_true {success_true}')
    print(f'success_false {success_false}')
    print(f'class_zero {class_zero}')
    print(f'class_nonzero {class_nonzero}')
    print(f'class_nan {class_nan}')
    print(f'success_true + success_false + class_zero {success_true + success_false + class_zero}')
    print(f'df len {len(df)}')
    print(f'class_zero + class_nonzero + class_nan {class_zero + class_nonzero + class_nan}')
    print(f'len(df) - class_zero - class_nan {len(df) - class_zero - class_nan }')

summarize_results(df)

# TEXT TOO SHORT

count = ((df['success'] == True) & (df['percent_human'] == 0) & (df['percent_ai'] == 0)).sum()
print("Number of rows where 'success' is True and 'percent_human' and 'percent_ai' are both 0:", count)
print ('')


# WHERE % HUMAN AND % AI ARE BOTH 0 THE WORD_COUNT WAS TOO SHORT
short_df = df[(df['success'] == True) & (df['percent_human'] == 0) & (df['percent_ai'] == 0)]
short_df

# 

data_df = df[(df['percent_human'] + df['percent_ai'] >= 99)]
data_df

# CALCULATE PERCENT
new_df = df[(df['percent_human'] + df['percent_ai'] >= 99)]
percentage = round((len(new_df) / len(df)) * 100, 1)
print("Percentage of rows where the sum of 'percent_human' and 'percent_ai' is 99 or greater:", percentage, "%")


def count_occurrences(df):
    result = df.groupby('kw').apply(lambda x: ((x['percent_human'] + x['percent_ai']) >= 99).sum()).reset_index(name='count')
    avg_values = df.groupby('kw').agg({'percent_human': [lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.min(),
                                                       lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.mean(),
                                                       lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.max()],
                                       'percent_ai': [lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.min(),
                                                      lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.mean(),
                                                      lambda x: np.nan if x.isnull().all() or (x == 0).all() else x.max()]}).reset_index()
    avg_values.columns = ['kw', '%_Human_Min', '%_Human_Mean', '%_Human_Max', '%_AI_Min', '%_AI_Mean', '%_AI_max']
    avg_values = avg_values.round({'%_Human_Min': 1, '%_Human_Mean': 1, '%_Human_Max': 1, '%_AI_Min': 1, '%_AI_Mean': 1, '%_AI_max': 1})
    result = pd.merge(result, avg_values, on='kw')
    return result

z = count_occurrences(df)
z

z.to_csv('preliminary_results.csv', index=False)

data_df.info()

data_df.to_csv('data/data_clean.csv', index=False)
